+++ src/main/java/group/msg/at/cloud/common/observability/rest/trace/RestTraceContainerRequestFilter.java
@@ -1,9 +1,11 @@
 package group.msg.at.cloud.common.observability.rest.trace;
 
+import org.eclipse.microprofile.config.ConfigProvider;
 import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.enterprise.context.Dependent;
 import javax.inject.Inject;
 import javax.ws.rs.container.ContainerRequestContext;
 import javax.ws.rs.container.ContainerRequestFilter;
@@ -12,11 +14,8 @@ import java.io.IOException;
 
 /**
  * {@code JAX-RS ContainerRequestFilter} which traces inbound requests from upstream services or frontends.
- * <p>
- * In contrast to ClientRequestFilters both CDI injection and automatic detection works well for ContainerRequestFilters
- * on all MicroProfile capable application servers.
- * </p>
  */
+@Dependent
 @Provider
 public class RestTraceContainerRequestFilter implements ContainerRequestFilter {
 
@@ -24,7 +23,7 @@ public class RestTraceContainerRequestFilter implements ContainerRequestFilter {
 
     @Inject
     @ConfigProperty(name = RestTraceConstants.ENABLED_CONFIG_KEY, defaultValue = RestTraceConstants.ENABLED_DEFAULT_VALUE)
-    boolean enabled;
+    Boolean enabled;
 
     @Inject
     RestTraceMessageBuilder messageBuilder;
@@ -33,12 +32,27 @@ public class RestTraceContainerRequestFilter implements ContainerRequestFilter {
     public void filter(ContainerRequestContext requestContext) throws IOException {
         if (shouldFilter(requestContext)) {
             StringBuilder traceMessage = new StringBuilder();
-            this.messageBuilder.build(traceMessage, requestContext);
+            getMessageBuilder().build(traceMessage, requestContext);
             LOGGER.info(traceMessage.toString());
         }
     }
 
     private boolean shouldFilter(ContainerRequestContext requestContext) {
-        return enabled && LOGGER.isInfoEnabled() && !requestContext.getUriInfo().getPath().contains("probes");
+        return isEnabled() && LOGGER.isInfoEnabled() && !requestContext.getUriInfo().getPath().contains("probes");
     }
+
+    private boolean isEnabled() {
+        if (enabled == null) {
+            ConfigProvider.getConfig().getOptionalValue(RestTraceConstants.ENABLED_CONFIG_KEY, Boolean.class).ifPresentOrElse(v -> enabled = v, () -> enabled = Boolean.FALSE);
+        }
+        return enabled;
+    }
+
+    private RestTraceMessageBuilder getMessageBuilder() {
+        if (this.messageBuilder == null) {
+            throw new IllegalStateException("Missing required reference to RestTraceMessageBuilder! Does your MicroProfile server supports CDI-injection in JAX-RS beans?");
+        }
+        return this.messageBuilder;
+    }
+
 }