+++ src/main/java/group/msg/at/cloud/common/observability/rest/trace/RestTraceContainerResponseFilter.java
@@ -1,5 +1,6 @@
 package group.msg.at.cloud.common.observability.rest.trace;
 
+import org.eclipse.microprofile.config.ConfigProvider;
 import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -25,7 +26,7 @@ public class RestTraceContainerResponseFilter implements ContainerResponseFilter
 
     @Inject
     @ConfigProperty(name = RestTraceConstants.ENABLED_CONFIG_KEY, defaultValue = RestTraceConstants.ENABLED_DEFAULT_VALUE)
-    boolean enabled;
+    Boolean enabled;
 
     @Inject
     RestTraceMessageBuilder messageBuilder;
@@ -34,13 +35,26 @@ public class RestTraceContainerResponseFilter implements ContainerResponseFilter
     public void filter(ContainerRequestContext requestContext, ContainerResponseContext responseContext) throws IOException {
         if (shouldFilter(requestContext, responseContext)) {
             StringBuilder traceMessage = new StringBuilder();
-            this.messageBuilder.build(traceMessage, requestContext, responseContext);
+            getMessageBuilder().build(traceMessage, requestContext, responseContext);
             LOGGER.info(traceMessage.toString());
         }
     }
 
     private boolean shouldFilter(ContainerRequestContext requestContext, ContainerResponseContext responseContext) {
-        return enabled && LOGGER.isInfoEnabled() && !requestContext.getUriInfo().getPath().contains("probes");
+        return isEnabled() && LOGGER.isInfoEnabled() && !requestContext.getUriInfo().getPath().contains("probes");
     }
 
+    private boolean isEnabled() {
+        if (enabled == null) {
+            ConfigProvider.getConfig().getOptionalValue(RestTraceConstants.ENABLED_CONFIG_KEY, Boolean.class).ifPresentOrElse(v -> enabled = v, () -> enabled = Boolean.FALSE);
+        }
+        return enabled;
+    }
+
+    private RestTraceMessageBuilder getMessageBuilder() {
+        if (this.messageBuilder == null) {
+            throw new IllegalStateException("Missing required reference to RestTraceMessageBuilder! Does your MicroProfile server supports CDI-injection in JAX-RS beans?");
+        }
+        return this.messageBuilder;
+    }
 }